All REST requests betterd + Postman workspace updated
Now almost all REST requests use @RequestParam instead of @RequestBody

Only POST request uses @RequesBody

+++ RESTbanking/src/main/java/be/uantwerpen/RESTbanking/Account.java
@@ -62,7 +62,7 @@ public class Account {
         this.password = password;
     }
 
-    public String deposit(long depositAmount) {
+    public String deposit(double depositAmount) {
         balance += depositAmount;
         return "amount deposited = " + depositAmount;
     }
